+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterAsyncGetTest.java
@@ -30,7 +30,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterAsyncGetTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterAsyncGetTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
