+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterAsyncSetTest.java
@@ -21,7 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterAsyncSetTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterAsyncSetTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
