+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterByteArrTransformerTest.java
@@ -5,9 +5,12 @@ import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.db.redis.RedisTransformStringAnalysisConfigurationFile;
+import cloud.filibuster.junit.configuration.examples.db.redis.RedisTransformBitInByteArrAnalysisConfigurationFile;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.sync.RedisCommands;
+import io.lettuce.core.codec.ByteArrayCodec;
+import io.lettuce.core.codec.RedisCodec;
+import io.lettuce.core.codec.StringCodec;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
@@ -15,45 +18,46 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
-import java.util.HashSet;
-import java.util.Set;
+import java.nio.charset.Charset;
+import java.util.ArrayList;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
 import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterStringTransformerTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterByteArrTransformerTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
-    static final String value = "example";
-    static StatefulRedisConnection<String, String> statefulRedisConnection;
+    static final byte[] value = "example".getBytes(Charset.defaultCharset());
+    static StatefulRedisConnection<String, byte[]> statefulRedisConnection;
     static String redisConnectionString;
-    private final static Set<String> testExceptionsThrown = new HashSet<>();
+    private final static ArrayList<String> testExceptionsThrown = new ArrayList<>();
 
     private static int numberOfTestExecutions = 0;
 
     @BeforeAll
     public static void primeCache() {
-        statefulRedisConnection = RedisClientService.getInstance().redisClient.connect();
+        statefulRedisConnection = RedisClientService.getInstance().redisClient.connect(RedisCodec.of(new StringCodec(), new ByteArrayCodec()));
         redisConnectionString = RedisClientService.getInstance().connectionString;
         statefulRedisConnection.sync().set(key, value);
     }
 
-    @DisplayName("Tests whether Redis sync interceptor can read from existing key - String transformer faults.")
+    @DisplayName("Tests whether Redis sync interceptor can read from existing key - Byte Array transformer faults.")
     @Order(1)
-    @TestWithFilibuster(analysisConfigurationFile = RedisTransformStringAnalysisConfigurationFile.class)
-    public void testRedisStringTransformation() {
+    @TestWithFilibuster(analysisConfigurationFile = RedisTransformBitInByteArrAnalysisConfigurationFile.class)
+    public void testRedisByteArrTransformation() {
         try {
             numberOfTestExecutions++;
 
-            StatefulRedisConnection<String, String> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
-            RedisCommands<String, String> myRedisCommands = myStatefulRedisConnection.sync();
-            String returnVal = myRedisCommands.get(key);
-            assertEquals(value, returnVal);
+            StatefulRedisConnection<String, byte[]> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
+            RedisCommands<String, byte[]> myRedisCommands = myStatefulRedisConnection.sync();
+            byte[] returnVal = myRedisCommands.get(key);
+            assertArrayEquals(value, returnVal);
             assertFalse(wasFaultInjected());
         } catch (Throwable t) {
             testExceptionsThrown.add(t.getMessage());
@@ -67,16 +71,17 @@ public class JUnitRedisFilibusterStringTransformerTest extends JUnitAnnotationBa
     @DisplayName("Verify correct number of test executions.")
     @Test
     @Order(2)
-    // 1 for the original test and +1 for each character manipulation in the string
+    // 1 for the original test and +1 for each bit in the byte array
     public void testNumExecutions() {
-        assertEquals(value.length() + 1, numberOfTestExecutions);
+        assertEquals(value.length * 8 + 1, numberOfTestExecutions);
     }
 
-    @DisplayName("Verify correct number of generated Filibuster tests.")
+    @DisplayName("Verify correct number of faults.")
     @Test
     @Order(3)
-    public void testNumExceptions() {
-        assertEquals(value.length(), testExceptionsThrown.size());
+    // 1 fault per bit per byte in the byte array
+    public void testNumFaults() {
+        assertEquals(value.length * 8, testExceptionsThrown.size());
     }
 
 }
