+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterByzantineStringTest.java
@@ -28,7 +28,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterByzantineStringTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterByzantineStringTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
