+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterFaultyExhaustiveExceptionAndStringTransformerTest.java
@@ -5,8 +5,7 @@ import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.db.redis.RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile;
-import io.lettuce.core.RedisCommandTimeoutException;
+import cloud.filibuster.functional.java.redis.transformers.configuration.FaultyRedisTransformStringAnalysisConfigurationFile;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.sync.RedisCommands;
 import org.junit.jupiter.api.BeforeAll;
@@ -16,9 +15,7 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
@@ -30,48 +27,43 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterSyncGetDefaultTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterFaultyExhaustiveExceptionAndStringTransformerTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
     static String redisConnectionString;
     private final static Set<String> testExceptionsThrown = new HashSet<>();
-
     private static int numberOfTestExecutions = 0;
 
-    private static final List<String> allowedExceptionMessages = new ArrayList<>();
-
     @BeforeAll
-    public static void primeCache() {
+    public static void beforeAll() {
         statefulRedisConnection = RedisClientService.getInstance().redisClient.connect();
         redisConnectionString = RedisClientService.getInstance().connectionString;
-        statefulRedisConnection.sync().set(key, value);
     }
 
-    static {
-        allowedExceptionMessages.add("Command timed out after 100 millisecond(s)");
-    }
-
-    @DisplayName("Tests whether Redis sync interceptor can read from existing key - Multiple fault injections")
+    @DisplayName("Tests whether Redis sync interceptor can read from existing key - String transformer faults with Faulty Accumulator.")
     @Order(1)
-    @TestWithFilibuster(analysisConfigurationFile = RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile.class)
-    public void testRedisSyncGetDefaultTests() {
+    @TestWithFilibuster(analysisConfigurationFile = FaultyRedisTransformStringAnalysisConfigurationFile.class)
+    public void testRedisStringBFIWithFaultyAccumulator() {
         try {
             numberOfTestExecutions++;
 
             StatefulRedisConnection<String, String> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
             RedisCommands<String, String> myRedisCommands = myStatefulRedisConnection.sync();
+            myRedisCommands.set(key, value);
+
             String returnVal = myRedisCommands.get(key);
             assertEquals(value, returnVal);
+
             assertFalse(wasFaultInjected());
         } catch (Throwable t) {
             testExceptionsThrown.add(t.getMessage());
 
-            assertTrue(wasFaultInjected(), "An exception was thrown although no fault was injected." + t);
+            assertTrue(wasFaultInjected(), "An exception was thrown although no fault was injected: " + t);
             assertThrows(FilibusterUnsupportedAPIException.class, () -> wasFaultInjectedOnService("io.lettuce.core.api.sync.RedisStringCommands"), "Expected FilibusterUnsupportedAPIException to be thrown: " + t);
-            assertTrue(wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/get"), "Fault was not injected on the Redis module." + t);
-            assertTrue(t instanceof RedisCommandTimeoutException, "Fault was not of the correct type: " + t);
-            assertTrue(allowedExceptionMessages.contains(t.getMessage()), "Unexpected fault message: " + t);
+            assertTrue(wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/get") ||
+                            wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/set"),
+                    "Fault was not injected on the expected Redis method: " + t);
         }
     }
 
@@ -79,13 +71,15 @@ public class JUnitRedisFilibusterSyncGetDefaultTest extends JUnitAnnotationBaseT
     @Test
     @Order(2)
     public void testNumExecutions() {
+        // 1 reference execution + 1 fault injections
         assertEquals(2, numberOfTestExecutions);
     }
 
-    @DisplayName("Verify correct number of generated Filibuster tests.")
+    @DisplayName("Verify correct number of unique injected faults.")
     @Test
     @Order(3)
     public void testNumExceptions() {
+        // Only 1 exception from mutating char 0 of string "example" to "X"
         assertEquals(1, testExceptionsThrown.size());
     }
