+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterStringTransformerTest.java
@@ -5,8 +5,7 @@ import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.db.redis.RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile;
-import io.lettuce.core.RedisCommandTimeoutException;
+import cloud.filibuster.junit.configuration.examples.db.redis.RedisTransformStringAnalysisConfigurationFile;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.sync.RedisCommands;
 import org.junit.jupiter.api.BeforeAll;
@@ -16,9 +15,7 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
@@ -26,12 +23,11 @@ import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterSyncGetNonExistingKeyTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterStringTransformerTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
@@ -40,8 +36,6 @@ public class JUnitRedisFilibusterSyncGetNonExistingKeyTest extends JUnitAnnotati
 
     private static int numberOfTestExecutions = 0;
 
-    private static final List<String> allowedExceptionMessages = new ArrayList<>();
-
     @BeforeAll
     public static void primeCache() {
         statefulRedisConnection = RedisClientService.getInstance().redisClient.connect();
@@ -49,21 +43,17 @@ public class JUnitRedisFilibusterSyncGetNonExistingKeyTest extends JUnitAnnotati
         statefulRedisConnection.sync().set(key, value);
     }
 
-    static {
-        allowedExceptionMessages.add("Command timed out after 100 millisecond(s)");
-    }
-
-    @DisplayName("Tests how Redis sync interceptor handles reading from non-existing key")
+    @DisplayName("Tests whether Redis sync interceptor can read from existing key - String transformer faults.")
     @Order(1)
-    @TestWithFilibuster(analysisConfigurationFile = RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile.class)
-    public void testRedisSyncGetNonExistingKey() {
+    @TestWithFilibuster(analysisConfigurationFile = RedisTransformStringAnalysisConfigurationFile.class)
+    public void testRedisStringTransformation() {
         try {
             numberOfTestExecutions++;
 
             StatefulRedisConnection<String, String> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
             RedisCommands<String, String> myRedisCommands = myStatefulRedisConnection.sync();
-            String returnVal = myRedisCommands.get("ThisKeyDoesNotExist");
-            assertNull(returnVal);
+            String returnVal = myRedisCommands.get(key);
+            assertEquals(value, returnVal);
             assertFalse(wasFaultInjected());
         } catch (Throwable t) {
             testExceptionsThrown.add(t.getMessage());
@@ -71,23 +61,22 @@ public class JUnitRedisFilibusterSyncGetNonExistingKeyTest extends JUnitAnnotati
             assertTrue(wasFaultInjected(), "An exception was thrown although no fault was injected: " + t);
             assertThrows(FilibusterUnsupportedAPIException.class, () -> wasFaultInjectedOnService("io.lettuce.core.api.sync.RedisStringCommands"), "Expected FilibusterUnsupportedAPIException to be thrown: " + t);
             assertTrue(wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/get"), "Fault was not injected on the expected Redis method: " + t);
-            assertTrue(t instanceof RedisCommandTimeoutException, "Fault was not of the correct type: " + t);
-            assertTrue(allowedExceptionMessages.contains(t.getMessage()), "Unexpected fault message: " + t);
         }
     }
 
     @DisplayName("Verify correct number of test executions.")
     @Test
     @Order(2)
+    // 1 for the original test and +1 for each character manipulation in the string
     public void testNumExecutions() {
-        assertEquals(2, numberOfTestExecutions);
+        assertEquals(value.length() + 1, numberOfTestExecutions);
     }
 
     @DisplayName("Verify correct number of generated Filibuster tests.")
     @Test
     @Order(3)
     public void testNumExceptions() {
-        assertEquals(1, testExceptionsThrown.size());
+        assertEquals(value.length(), testExceptionsThrown.size());
     }
 
 }
