+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterSyncGetDefaultTest.java
@@ -30,7 +30,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterSyncGetTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterSyncGetDefaultTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
@@ -52,10 +52,10 @@ public class JUnitRedisFilibusterSyncGetTest extends JUnitAnnotationBaseTest {
         allowedExceptionMessages.add("Command timed out after 100 millisecond(s)");
     }
 
-    @DisplayName("Tests whether Redis sync interceptor can read from existing key - Single fault injection")
+    @DisplayName("Tests whether Redis sync interceptor can read from existing key - Multiple fault injections")
     @Order(1)
     @TestWithFilibuster(analysisConfigurationFile = RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile.class)
-    public void testRedisSyncGet() {
+    public void testRedisSyncGetDefaultTests() {
         try {
             numberOfTestExecutions++;
 
@@ -67,9 +67,9 @@ public class JUnitRedisFilibusterSyncGetTest extends JUnitAnnotationBaseTest {
         } catch (Throwable t) {
             testExceptionsThrown.add(t.getMessage());
 
-            assertTrue(wasFaultInjected(), "An exception was thrown although no fault was injected: " + t);
+            assertTrue(wasFaultInjected(), "An exception was thrown although no fault was injected." + t);
             assertThrows(FilibusterUnsupportedAPIException.class, () -> wasFaultInjectedOnService("io.lettuce.core.api.sync.RedisStringCommands"), "Expected FilibusterUnsupportedAPIException to be thrown: " + t);
-            assertTrue(wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/get"), "Fault was not injected on the expected Redis method: " + t);
+            assertTrue(wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/get"), "Fault was not injected on the Redis module." + t);
             assertTrue(t instanceof RedisCommandTimeoutException, "Fault was not of the correct type: " + t);
             assertTrue(allowedExceptionMessages.contains(t.getMessage()), "Unexpected fault message: " + t);
         }
