+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterSyncGetNonExistingKeyTest.java
@@ -1,11 +1,12 @@
-package cloud.filibuster.functional.java.redis.scenarios;
+package cloud.filibuster.functional.java.redis;
 
 import cloud.filibuster.exceptions.filibuster.FilibusterUnsupportedAPIException;
 import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.functional.java.redis.transformers.configuration.FaultyHasNextFalseRedisTransformStringAnalysisConfigurationFile;
+import cloud.filibuster.junit.configuration.examples.db.redis.RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile;
+import io.lettuce.core.RedisCommandTimeoutException;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.sync.RedisCommands;
 import org.junit.jupiter.api.BeforeAll;
@@ -15,7 +16,9 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
@@ -23,47 +26,53 @@ import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterFaultyHasNextFalseStringTransformerTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterSyncGetNonExistingKeyTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
     static String redisConnectionString;
     private final static Set<String> testExceptionsThrown = new HashSet<>();
+
     private static int numberOfTestExecutions = 0;
 
+    private static final List<String> allowedExceptionMessages = new ArrayList<>();
+
     @BeforeAll
-    public static void beforeAll() {
+    public static void primeCache() {
         statefulRedisConnection = RedisClientService.getInstance().redisClient.connect();
         redisConnectionString = RedisClientService.getInstance().connectionString;
+        statefulRedisConnection.sync().set(key, value);
     }
 
-    @DisplayName("Tests the scenario where the accumulator is faulty. In this case, the accumulator hasNext() method always returns false")
+    static {
+        allowedExceptionMessages.add("Command timed out after 100 millisecond(s)");
+    }
+
+    @DisplayName("Tests how Redis sync interceptor handles reading from non-existing key")
     @Order(1)
-    @TestWithFilibuster(analysisConfigurationFile = FaultyHasNextFalseRedisTransformStringAnalysisConfigurationFile.class)
-    public void testRedisStringBFIWithFaultyAccumulator() {
+    @TestWithFilibuster(analysisConfigurationFile = RedisSingleFaultCommandTimeoutExceptionAnalysisConfigurationFile.class)
+    public void testRedisSyncGetNonExistingKey() {
         try {
             numberOfTestExecutions++;
 
             StatefulRedisConnection<String, String> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
             RedisCommands<String, String> myRedisCommands = myStatefulRedisConnection.sync();
-            myRedisCommands.set(key, value);
-
-            String returnVal = myRedisCommands.get(key);
-            assertEquals(value, returnVal);
-
+            String returnVal = myRedisCommands.get("ThisKeyDoesNotExist");
+            assertNull(returnVal);
             assertFalse(wasFaultInjected());
         } catch (Throwable t) {
             testExceptionsThrown.add(t.getMessage());
 
             assertTrue(wasFaultInjected(), "An exception was thrown although no fault was injected: " + t);
             assertThrows(FilibusterUnsupportedAPIException.class, () -> wasFaultInjectedOnService("io.lettuce.core.api.sync.RedisStringCommands"), "Expected FilibusterUnsupportedAPIException to be thrown: " + t);
-            assertTrue(wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/get") ||
-                            wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/set"),
-                    "Fault was not injected on the expected Redis method: " + t);
+            assertTrue(wasFaultInjectedOnJavaClassAndMethod("io.lettuce.core.api.sync.RedisStringCommands/get"), "Fault was not injected on the expected Redis method: " + t);
+            assertTrue(t instanceof RedisCommandTimeoutException, "Fault was not of the correct type: " + t);
+            assertTrue(allowedExceptionMessages.contains(t.getMessage()), "Unexpected fault message: " + t);
         }
     }
 
@@ -71,15 +80,13 @@ public class JUnitRedisFilibusterFaultyHasNextFalseStringTransformerTest extends
     @Test
     @Order(2)
     public void testNumExecutions() {
-        // 1 reference execution + 1 fault injection
         assertEquals(2, numberOfTestExecutions);
     }
 
-    @DisplayName("Verify correct number of unique injected faults.")
+    @DisplayName("Verify correct number of generated Filibuster tests.")
     @Test
     @Order(3)
     public void testNumExceptions() {
-        // Only 1 exception from mutating char 0 of string "example" to "X"
         assertEquals(1, testExceptionsThrown.size());
     }
