+++ src/test/java/cloud/filibuster/functional/java/redis/RedisFilibusterTrackedMethodsExceptionsTest.java
@@ -37,7 +37,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterTrackedMethodsExceptionsTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterTrackedMethodsExceptionsTest extends JUnitAnnotationBaseTest {
     static final String key = "test";
     static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
