+++ src/test/java/cloud/filibuster/functional/java/redis/scenarios/FaultyRedisFaultFreeGRPCTest.java
@@ -8,7 +8,7 @@ import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInte
 import cloud.filibuster.instrumentation.libraries.grpc.FilibusterClientInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.db.redis.GrpcAndRedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile;
+import cloud.filibuster.junit.configuration.examples.db.redis.RedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile;
 import io.grpc.Channel;
 import io.grpc.ClientInterceptor;
 import io.grpc.ClientInterceptors;
@@ -38,20 +38,19 @@ import static cloud.filibuster.integration.instrumentation.TestHelper.startHello
 import static cloud.filibuster.integration.instrumentation.TestHelper.stopHelloServerAndWaitUntilUnavailable;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
-import static cloud.filibuster.junit.assertions.GrpcAssertions.wasFaultInjectedOnMethod;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFaultyRedisFaultyGRPCTest extends JUnitAnnotationBaseTest {
+public class FaultyRedisFaultFreeGRPCTest extends JUnitAnnotationBaseTest {
     static StatefulRedisConnection<String, String> statefulRedisConnection;
     static String redisConnectionString;
-    private static final Logger logger = Logger.getLogger(JUnitFaultyRedisFaultyGRPCTest.class.getName());
+    private static final Logger logger = Logger.getLogger(FaultyRedisFaultFreeGRPCTest.class.getName());
     private static final ArrayList<String> keys = new ArrayList<>();
     private static final ArrayList<String> values = new ArrayList<>();
-    private static String name;
     private static int numberOfExecution = 0;
     private static final HashSet<String> actualFaultMessages = new HashSet<>();
+    private static String name;
     private static final Random rand = new Random(0);
 
     @BeforeAll
@@ -61,7 +60,7 @@ public class JUnitFaultyRedisFaultyGRPCTest extends JUnitAnnotationBaseTest {
 
         // Generate random keys and values and add them to Redis
         RedisCommands<String, String> redisCommands = statefulRedisConnection.sync();
-        for (int i = 0; i < 2; i++) {
+        for (int i = 0; i < 3; i++) {
             String key = getRandomString();
             String value = getRandomString();
             keys.add(key);
@@ -79,29 +78,32 @@ public class JUnitFaultyRedisFaultyGRPCTest extends JUnitAnnotationBaseTest {
         stopHelloServerAndWaitUntilUnavailable();
     }
 
-    @DisplayName("Tests the scenario where faults are injected in both Redis and the GRPC client. GRPC calls are issued before and after the Redis call")
+    @DisplayName("Tests the scenario where faults are injected in Redis, while the GRPC client before and after the Redis call " +
+            "is fault-free")
     @Order(1)
     @TestWithFilibuster(
-            analysisConfigurationFile = GrpcAndRedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile.class,
-            maxIterations = 300
+            analysisConfigurationFile = RedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile.class,
+            maxIterations = 200
     )
     public void testRedisSync() {
         numberOfExecution++;
 
         // Send GRPC request
-        sayHelloAndAssert(name);
+        Hello.HelloReply helloReply = sayHello(name);
+        assertEquals(String.format("Hello, %s!!", name), helloReply.getMessage());
 
         // Prepare Redis interceptor
         StatefulRedisConnection<String, String> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
         RedisCommands<String, String> myRedisCommands = myStatefulRedisConnection.sync();
 
         // Get key from Redis and assert correct value
-        for (int i = 0; i < 2; i++) {
+        for (int i = 0; i < 3; i++) {
             getFromRedisAndAssert(myRedisCommands, keys.get(i), values.get(i));
         }
 
         // Send GRPC request
-        sayHelloAndAssert(name);
+        helloReply = sayHello(name);
+        assertEquals(String.format("Hello, %s!!", name), helloReply.getMessage());
     }
 
     @DisplayName("Assert correct number of test executions")
@@ -110,23 +112,18 @@ public class JUnitFaultyRedisFaultyGRPCTest extends JUnitAnnotationBaseTest {
     public void testNumberOfExecutions() {
         // Each Redis get call leads to 5 executions: 1 fault-free execution, 2 transformer faults (one for each char),
         // 1 byzantine execution (injecting null) and 1 exception execution (injecting RedisCommandTimeoutException)
-        // Redis get is called 2 times, leading to 5^2 = 25 executions
-        // Each GRPC call leads to 2 execution: One is fault free, while the other throws a StatusRuntimeException.
-        // There are two GRPC calls. Therefore, we expect 25 + 2 * 2 = 100 executions
-        assertEquals(100, numberOfExecution);
+        // Redis get is called 3 times, leading to 5^3 = 125 executions
+        assertEquals(125, numberOfExecution);
     }
 
     @DisplayName("Assert number of faults")
     @Order(3)
     @Test
     public void testNumberOfFaults() {
-        // For each of the 2 Redis get call, we inject 4 faults: 2 transformer faults (one for each char), 1 byzantine
-        // fault and 1 exception. The error message of the exception is the same for all Redis get calls.
-        // Therefore, we expect 1 + 2 * 3 = 7 Redis fault messages
-        // Additionally, we have 2 GRPC calls. Each can throw an UNAVAILABLE exception.
-        // The exception message is the same for both GRPC calls.
-        // Therefore, we expect 7 + 1 = 8 fault messages
-        assertEquals(8, actualFaultMessages.size());
+        // For each of the 3 Redis get call, we inject 4 faults: 2 transformer faults (one for each char), 1 byzantine fault and 1 exception.
+        // The error message of the exception is the same for all Redis get calls.
+        // Therefore, we expect 1 + 3 * 3 = 10 unique fault messages
+        assertEquals(10, actualFaultMessages.size());
     }
 
     @DisplayName("Assert correct fault messages")
@@ -136,12 +133,10 @@ public class JUnitFaultyRedisFaultyGRPCTest extends JUnitAnnotationBaseTest {
         List<String> transformerFaults = getMatchesInFaultMessages("expected: <..> but was: <..>");
         List<String> nullFaults = getMatchesInFaultMessages("expected: <..> but was: <null>");
         List<String> timeoutException = getMatchesInFaultMessages("Command timed out after 100 millisecond\\(s\\)");
-        List<String> grpcException = getMatchesInFaultMessages("UNAVAILABLE");
 
-        assertEquals(4, transformerFaults.size());
-        assertEquals(2, nullFaults.size());
+        assertEquals(6, transformerFaults.size());
+        assertEquals(3, nullFaults.size());
         assertEquals(1, timeoutException.size());
-        assertEquals(1, grpcException.size());
     }
 
     private static List<String> getMatchesInFaultMessages(String regex) {
@@ -153,26 +148,18 @@ public class JUnitFaultyRedisFaultyGRPCTest extends JUnitAnnotationBaseTest {
                 .collect(Collectors.toList());
     }
 
-    private static void sayHelloAndAssert(String name) {
-        try {
-            ManagedChannel helloChannel = ManagedChannelBuilder.forAddress(Networking.getHost("hello"), Networking.getPort("hello")).usePlaintext().build();
+    private static Hello.HelloReply sayHello(String name) {
+        ManagedChannel helloChannel = ManagedChannelBuilder.forAddress(Networking.getHost("hello"), Networking.getPort("hello")).usePlaintext().build();
 
-            ClientInterceptor clientInterceptor = new FilibusterClientInterceptor("hello");
-            Channel channel = ClientInterceptors.intercept(helloChannel, clientInterceptor);
+        ClientInterceptor clientInterceptor = new FilibusterClientInterceptor("hello");
+        Channel channel = ClientInterceptors.intercept(helloChannel, clientInterceptor);
 
-            HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(channel);
-            Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName(name).build();
-            Hello.HelloReply helloReply = blockingStub.hello(request);
+        HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(channel);
+        Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName(name).build();
+        Hello.HelloReply helloReply = blockingStub.hello(request);
 
-            helloChannel.shutdown();
-
-            assertEquals(String.format("Hello, %s!!", name), helloReply.getMessage());
-        } catch (Throwable e) {
-            logger.log(Level.INFO, "getFromRedis threw an exception: " + e);
-            actualFaultMessages.add(e.getMessage());
-            assertTrue(wasFaultInjected());
-            assertTrue(wasFaultInjectedOnMethod(HelloServiceGrpc.getHelloMethod()));
-        }
+        helloChannel.shutdown();
+        return helloReply;
     }
 
     private static void getFromRedisAndAssert(RedisCommands<String, String> redisCommand, String key, String value) {
@@ -189,4 +176,5 @@ public class JUnitFaultyRedisFaultyGRPCTest extends JUnitAnnotationBaseTest {
 
     private static String getRandomString() {
         return String.valueOf(rand.nextInt(90) + 10);
-    }}
+    }
+}
