+++ src/test/java/cloud/filibuster/functional/java/redis/scenarios/FaultyRedisFaultyGRPCWithDataNondeterminismAndSuppressCombinationsTest.java
@@ -8,7 +8,7 @@ import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInte
 import cloud.filibuster.instrumentation.libraries.grpc.FilibusterClientInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.db.redis.RedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile;
+import cloud.filibuster.junit.configuration.examples.db.redis.GrpcAndRedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile;
 import io.grpc.Channel;
 import io.grpc.ClientInterceptor;
 import io.grpc.ClientInterceptors;
@@ -38,14 +38,15 @@ import static cloud.filibuster.integration.instrumentation.TestHelper.startHello
 import static cloud.filibuster.integration.instrumentation.TestHelper.stopHelloServerAndWaitUntilUnavailable;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
+import static cloud.filibuster.junit.assertions.GrpcAssertions.wasFaultInjectedOnMethod;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombinationsTest extends JUnitAnnotationBaseTest {
+public class FaultyRedisFaultyGRPCWithDataNondeterminismAndSuppressCombinationsTest extends JUnitAnnotationBaseTest {
     static StatefulRedisConnection<String, String> statefulRedisConnection;
     static String redisConnectionString;
-    private static final Logger logger = Logger.getLogger(JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombinationsTest.class.getName());
+    private static final Logger logger = Logger.getLogger(FaultyRedisFaultyGRPCWithDataNondeterminismAndSuppressCombinationsTest.class.getName());
     private static final ArrayList<String> keys = new ArrayList<>();
     private static final ArrayList<String> values = new ArrayList<>();
     private static int numberOfExecution = 0;
@@ -75,12 +76,12 @@ public class JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombi
         stopHelloServerAndWaitUntilUnavailable();
     }
 
-    @DisplayName("Tests the scenario where faults are injected in Redis, while the GRPC client before and after the Redis call " +
-            "is fault-free. Data nondeterminism and suppressCombinations are enabled.")
+    @DisplayName("Tests the scenario where faults are injected in both Redis and the GRPC client. GRPC calls are issued before and after the Redis call. " +
+            "Both dataNondeterminism and suppressCombinations are true.")
     @Order(1)
     @TestWithFilibuster(
-            analysisConfigurationFile = RedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile.class,
-            maxIterations = 20,
+            analysisConfigurationFile = GrpcAndRedisStringExceptionAndTransformerAndByzantineAnalysisConfigurationFile.class,
+            maxIterations = 30,
             dataNondeterminism = true,
             suppressCombinations = true
     )
@@ -88,9 +89,7 @@ public class JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombi
         numberOfExecution++;
 
         // Send GRPC request with random name
-        String name = getRandomString();
-        Hello.HelloReply helloReply = sayHello(name);
-        assertEquals(String.format("Hello, %s!!", name), helloReply.getMessage());
+        sayHelloAndAssert(getRandomString());
 
         // Prepare Redis interceptor
         StatefulRedisConnection<String, String> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
@@ -102,9 +101,7 @@ public class JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombi
         }
 
         // Send GRPC request with random name
-        name = getRandomString();
-        helloReply = sayHello(name);
-        assertEquals(String.format("Hello, %s!!", name), helloReply.getMessage());
+        sayHelloAndAssert(getRandomString());
     }
 
     @DisplayName("Assert correct number of test executions")
@@ -113,18 +110,23 @@ public class JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombi
     public void testNumberOfExecutions() {
         // We inject 4 faults per Redis get call: 2 transformer faults (one for each char),
         // 1 byzantine execution (injecting null) and 1 exception execution (injecting RedisCommandTimeoutException)
-        // Since suppressCombinations is true and Redis get is called 3 times, this leads to 1 reference execution + 4*3 = 13 executions
-        assertEquals(13, numberOfExecution);
+        // Per GRPC call, we inject one UNAVAILABLE exception.
+        // Since suppressCombinations is true, Redis get is called 3 times and GRPC calls are issued twice,
+        // this leads to 1 reference execution + 4*3 + 2 = 15 executions
+        assertEquals(15, numberOfExecution);
     }
 
     @DisplayName("Assert number of faults")
     @Order(3)
     @Test
     public void testNumberOfFaults() {
-        // For each of the 3 Redis get call, we inject 4 faults: 2 transformer faults (one for each char), 1 byzantine fault and 1 exception.
-        // The error message of the exception is the same for all Redis get calls.
-        // Therefore, we expect 1 + 3 * 3 = 10 unique fault messages
-        assertEquals(10, actualFaultMessages.size());
+        // For each of the 2 Redis get call, we inject 4 faults: 2 transformer faults (one for each char), 1 byzantine
+        // fault and 1 exception. The error message of the exception is the same for all Redis get calls.
+        // Therefore, we expect 1 + 3 * 3 = 10 Redis fault messages
+        // Additionally, we have 2 GRPC calls. Each can throw an UNAVAILABLE exception.
+        // The exception message is the same for both GRPC calls.
+        // Therefore, we expect 10 + 1 = 11 fault messages
+        assertEquals(11, actualFaultMessages.size());
     }
 
     @DisplayName("Assert correct fault messages")
@@ -134,33 +136,34 @@ public class JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombi
         List<String> transformerFaults = getMatchesInFaultMessages("expected: <..> but was: <..>");
         List<String> nullFaults = getMatchesInFaultMessages("expected: <..> but was: <null>");
         List<String> timeoutException = getMatchesInFaultMessages("Command timed out after 100 millisecond\\(s\\)");
+        List<String> grpcException = getMatchesInFaultMessages("UNAVAILABLE");
 
         assertEquals(6, transformerFaults.size());
         assertEquals(3, nullFaults.size());
         assertEquals(1, timeoutException.size());
+        assertEquals(1, grpcException.size());
     }
 
-    private static List<String> getMatchesInFaultMessages(String regex) {
-        Pattern pattern = Pattern.compile(regex);
+    private static void sayHelloAndAssert(String name) {
+        try {
+            ManagedChannel helloChannel = ManagedChannelBuilder.forAddress(Networking.getHost("hello"), Networking.getPort("hello")).usePlaintext().build();
 
-        return actualFaultMessages
-                .stream()
-                .filter(e -> pattern.matcher(e).matches())
-                .collect(Collectors.toList());
-    }
+            ClientInterceptor clientInterceptor = new FilibusterClientInterceptor("hello");
+            Channel channel = ClientInterceptors.intercept(helloChannel, clientInterceptor);
 
-    private static Hello.HelloReply sayHello(String name) {
-        ManagedChannel helloChannel = ManagedChannelBuilder.forAddress(Networking.getHost("hello"), Networking.getPort("hello")).usePlaintext().build();
+            HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(channel);
+            Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName(name).build();
+            Hello.HelloReply helloReply = blockingStub.hello(request);
 
-        ClientInterceptor clientInterceptor = new FilibusterClientInterceptor("hello");
-        Channel channel = ClientInterceptors.intercept(helloChannel, clientInterceptor);
+            helloChannel.shutdown();
 
-        HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(channel);
-        Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName(name).build();
-        Hello.HelloReply helloReply = blockingStub.hello(request);
-
-        helloChannel.shutdown();
-        return helloReply;
+            assertEquals(String.format("Hello, %s!!", name), helloReply.getMessage());
+        } catch (Throwable e) {
+            logger.log(Level.INFO, "getFromRedis threw an exception: " + e);
+            actualFaultMessages.add(e.getMessage());
+            assertTrue(wasFaultInjected());
+            assertTrue(wasFaultInjectedOnMethod(HelloServiceGrpc.getHelloMethod()));
+        }
     }
 
     private static void getFromRedisAndAssert(RedisCommands<String, String> redisCommand, String key, String value) {
@@ -175,7 +178,15 @@ public class JUnitFaultyRedisFaultFreeGRPCWithDataNondeterminismAndSuppressCombi
         }
     }
 
+    private static List<String> getMatchesInFaultMessages(String regex) {
+        Pattern pattern = Pattern.compile(regex);
+
+        return actualFaultMessages
+                .stream()
+                .filter(e -> pattern.matcher(e).matches())
+                .collect(Collectors.toList());
+    }
+
     private static String getRandomString() {
         return String.valueOf(rand.nextInt(90) + 10);
-    }
-}
+    }}
