+++ src/test/java/cloud/filibuster/functional/java/redis/scenarios/RedisFilibusterFaultyHasNextFalseStringTransformerTest.java
@@ -1,11 +1,11 @@
-package cloud.filibuster.functional.java.redis;
+package cloud.filibuster.functional.java.redis.scenarios;
 
 import cloud.filibuster.exceptions.filibuster.FilibusterUnsupportedAPIException;
 import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.RedisClientService;
 import cloud.filibuster.junit.TestWithFilibuster;
-import cloud.filibuster.junit.configuration.examples.db.redis.RedisExhaustiveExceptionAndTransformerAnalysisConfigurationFile;
+import cloud.filibuster.functional.java.redis.transformers.configuration.FaultyHasNextFalseRedisTransformStringAnalysisConfigurationFile;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.api.sync.RedisCommands;
 import org.junit.jupiter.api.BeforeAll;
@@ -17,24 +17,22 @@ import org.junit.jupiter.api.TestMethodOrder;
 
 import java.util.HashSet;
 import java.util.Set;
-import java.util.UUID;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
 import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitRedisFilibusterExhaustiveExceptionAndStringTransformerDataNonDeterminismTest extends JUnitAnnotationBaseTest {
+public class RedisFilibusterFaultyHasNextFalseStringTransformerTest extends JUnitAnnotationBaseTest {
+    static final String key = "test";
+    static final String value = "example";
     static StatefulRedisConnection<String, String> statefulRedisConnection;
     static String redisConnectionString;
     private final static Set<String> testExceptionsThrown = new HashSet<>();
-    private final static int keyLength = 5;
-    private final static int valueLength = 10;
     private static int numberOfTestExecutions = 0;
 
     @BeforeAll
@@ -43,35 +41,19 @@ public class JUnitRedisFilibusterExhaustiveExceptionAndStringTransformerDataNonD
         redisConnectionString = RedisClientService.getInstance().connectionString;
     }
 
-    @DisplayName("Tests whether Redis sync interceptor can read from existing key - Exhaustive Exception and String transformer faults.")
+    @DisplayName("Tests the scenario where the accumulator is faulty. In this case, the accumulator hasNext() method always returns false")
     @Order(1)
-    @TestWithFilibuster(
-            analysisConfigurationFile = RedisExhaustiveExceptionAndTransformerAnalysisConfigurationFile.class,
-            dataNondeterminism = true
-    )
-    public void testRedisStringBFIAndExceptionInjection() {
+    @TestWithFilibuster(analysisConfigurationFile = FaultyHasNextFalseRedisTransformStringAnalysisConfigurationFile.class)
+    public void testRedisStringBFIWithFaultyAccumulator() {
         try {
             numberOfTestExecutions++;
 
-            String key = generateRandomString(keyLength);
-            String value = generateRandomString(valueLength);
-
             StatefulRedisConnection<String, String> myStatefulRedisConnection = DynamicProxyInterceptor.createInterceptor(statefulRedisConnection, redisConnectionString);
             RedisCommands<String, String> myRedisCommands = myStatefulRedisConnection.sync();
-
             myRedisCommands.set(key, value);
+
             String returnVal = myRedisCommands.get(key);
             assertEquals(value, returnVal);
-
-            myRedisCommands.set(value, key);
-            returnVal = myRedisCommands.get(value);
-            assertEquals(key, returnVal);
-
-            assertNull(myRedisCommands.get("NonexistentKey"));
-
-            myRedisCommands.set(key, "");
-            returnVal = myRedisCommands.get(key);
-            assertEquals("", returnVal);
 
             assertFalse(wasFaultInjected());
         } catch (Throwable t) {
@@ -89,23 +71,16 @@ public class JUnitRedisFilibusterExhaustiveExceptionAndStringTransformerDataNonD
     @Test
     @Order(2)
     public void testNumExecutions() {
-        // Reference execution + 3 RedisCommandTimeoutExceptions injected on set + 4 RedisCommandTimeoutExceptions injected on get
-        // + transformer faults on key + transformer faults on test (4 chars)
-        assertEquals(8 + keyLength + valueLength, numberOfTestExecutions);
+        // 1 reference execution + 1 fault injection
+        assertEquals(2, numberOfTestExecutions);
     }
 
     @DisplayName("Verify correct number of unique injected faults.")
     @Test
     @Order(3)
     public void testNumExceptions() {
-        // RedisCommandTimeoutException injected on set/get
-        // + transformer faults on key + transformer faults on value
-        assertEquals(1 + keyLength + valueLength, testExceptionsThrown.size());
-    }
-
-    private static String generateRandomString(int length) {
-        String uuid = UUID.randomUUID().toString();
-        return uuid.substring(0, length);
+        // Only 1 exception from mutating char 0 of string "example" to "X"
+        assertEquals(1, testExceptionsThrown.size());
     }
 
 }
